ref(ddm): restructure ddm dir
