+++ static/app/views/ddm/constants.tsx
@@ -1,0 +1,3 @@
+export const DDM_CHART_GROUP = 'ddm_chart_group';
+
+export const MIN_WIDGET_WIDTH = 400;
