+++ static/app/views/ddm/queryBuilder.tsx
@@ -23,11 +23,11 @@ import useApi from 'sentry/utils/useApi';
 import useKeyPress from 'sentry/utils/useKeyPress';
 import useOrganization from 'sentry/utils/useOrganization';
 import usePageFilters from 'sentry/utils/usePageFilters';
-import {MetricWidgetProps} from 'sentry/views/ddm/metricWidget';
+import {MetricWidgetProps} from 'sentry/views/ddm/widget';
 
 type QueryBuilderProps = {
   displayType: MetricDisplayType; // TODO(ddm): move display type out of the query builder
-  metricsQuery: MetricsQuery;
+  metricsQuery: Pick<MetricsQuery, 'mri' | 'op' | 'query' | 'groupBy'>;
   onChange: (data: Partial<MetricWidgetProps>) => void;
   projects: number[];
   powerUserMode?: boolean;
